feat(PR-611): redesign chat + styling & add functionality to right sidebar
