+++ src/app/ui-components/sidebar-right/sidebar-right.component.css
@@ -68,7 +68,31 @@
     content: none;
 }
 
-.dropdown-item:hover {
-    color: #e10072;
+.dropdown-item {
     background-color: transparent;
+}
+
+.item {
+    max-height: 24px;
+}
+
+.dropdown-item:hover label,
+.dropdown-item:hover .item {
+    filter: invert(80%) sepia(50%) saturate(2000%) hue-rotate(300deg)
+        brightness(142%) contrast(110%)
+        drop-shadow(0px 0px 10px rgba(149, 27, 129, 1));
+}
+
+.off {
+    background-color: #0094df;
+}
+
+.on,
+.off:hover {
+    background-color: #e10072;
+}
+
+.onOff {
+    border-radius: 5px;
+    min-height: 30px;
 }
