+++ src/app/voice-assistant/voice-assistant-chat/chat-window/chat-window.component.html
@@ -15,16 +15,6 @@
     <div class="col text-left">
         <span class="align-middle"> Chats </span>
     </div>
-    <!-- <div class="col-8 d-flex justify-content-end">
-        <button
-            id="export-button"
-            class="btn btn-regular"
-            [ngClass]="{'disabled-button': !this.chat}"
-            (click)="this.exportChat()"
-            [disabled]="!this.chat">
-            Export chat
-        </button>
-    </div> -->
 </div>
 <div class="row mt-2 content-window pb-2">
     <div class="col-12">
