+++ src/app/voice-assistant/voice-assistant.component.html
@@ -1,22 +1,4 @@
-<div class="col">
-    <div class="row mt-5">
-        <div class="col-auto text-right">
-            <span class="align-middle"> Off </span>
-        </div>
-        <div class="col-auto pr-0 pl-0">
-            <button
-                class="toggle-button"
-                (click)="toggleVoiceAssistant()"
-                id="toggle-voice-assistant">
-                <img
-                    [src]="imgSrc"
-                    alt="Toggle for turning the Voice Assistant on and off" />
-            </button>
-        </div>
-        <div class="col-auto text-left">
-            <span class="align-middle"> On </span>
-        </div>
-    </div>
+<div class="col mt-4">
     <div class="row">
         <div class="col">
             <app-voice-assistant-nav
